requirements.txt
py-cord[voice]
discord-pretty-help
PyNaCl==1.4.0
async-timeout==3.0.1
asyncio==3.4.3
pillow
requests
youtube_dl
gmplot